% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{fuel}
\alias{fuel}
\title{Petrol prices in New South Wales, Australia, 2016 - 2017.}
\format{A data frame with 1042318 rows and 9 variables:
\describe{
  \item{ServiceStationName}{'Name' of the station, note there are fewer unique names than addresses.}
  \item{Address}{Physical address of the station.}
  \item{Suburb}{Town, locality, suburb}
  \item{Postcode}{NSW postal code}
  \item{Brand}{Caltex, BP, etc.}
  \item{FuelCode}{Type of fuel: P98, P95, U91, etc. (reporting frequency varies)}
  \item{Price}{Price in AUD cents}
  \item{DateTime}{Date and time of price report (POSIXct)}
  \item{Date}{Date (Date)}
}}
\source{
Made possible by the FuelCheck service in NSW (\url{https://www.fuelcheck.nsw.gov.au/app}), this dataset was based on chunked data made available at \url{https://data.nsw.gov.au/data/dataset/fuel-check}
}
\usage{
fuel
}
\description{
Dataset of petrol prices for ca. 2000 service stations across New South Wales. Includes observations for 11 fuel types, 20 service station brands. Reporting frequency is often daily (for metropolitan stations), but less frequent for more remote stations.
}
\keyword{datasets}
