% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuelpricensw-package.r
\docType{package}
\name{fuelpricensw}
\alias{fuelpricensw}
\alias{fuelpricensw-package}
\title{Fuel prices in New South Wales}
\description{
Near-daily fuel prices for nearly all service stations in New South Wales, between August 2016 and October 201
}
